+++ CHANGELOG.md
@@ -7,10 +7,12 @@ v4.0.0 (by Gecko) - ??/??/????
 - UTF-8 support for all strings
 - `Update` is deprecated, use `EngineUpdate` instead
 - `TrisRendered` was removed
+- `ViewerRenderEx` was removed
 - New function `ViewerRenderObject`
 - `ViewerGetPickedObjectsList` now requires user-provied PickList
 - New functions `PickListCreate`, `PickListClear`, `PickListGetCount`, `PickListGetHit`
 - New constants for `ViewerSetAntiAliasing`: `aa6x = 6`, `aa8x = 7`, `aa16x = 8`, `csa8x = 9`, `csa8xHQ = 10`, `csa16x = 11`, `csa16xHQ = 12`
+- `ViewerRenderToFile` now supports PNG, `ViewerRenderToFilePNG` was removed
 - `CameraZoomAll` now requires a Viewer as a second parameter
 - New function `MaterialLibraryGetTextureByName`
 - `SpriteNoZWrite` was removed, use `MaterialSetDepthWrite` instead